+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Celestial Bloodline now works correctly with Metamagic Rager
 * Adjusted Magical Tail Feats @Balkoth
 * Kitsune polymorph is now correctly considered a polymorph
+* Updated string handling to support general localization (will need help from speakers to do actual localization work)
 
 ## Version 1.9.6b
 * Aspect of Omox DR now stacks globaly